refactor: moved data classes
