+++ cognite_toolkit/_api/modules_api.py
@@ -15,9 +15,9 @@ from cognite_toolkit._api.data_classes import _DUMMY_ENVIRONMENT, ModuleMeta, Mo
 from cognite_toolkit._cdf import Common, clean, deploy
 from cognite_toolkit._cdf_tk.commands.build import BuildCommand
 from cognite_toolkit._cdf_tk.constants import COGNITE_MODULES, COGNITE_MODULES_PATH
+from cognite_toolkit._cdf_tk.data_classes import BuildConfigYAML, Environment, InitConfigYAML, SystemYAML
 from cognite_toolkit._cdf_tk.loaders import ResourceTypes
 from cognite_toolkit._cdf_tk.templates import iterate_modules
-from cognite_toolkit._cdf_tk.templates.data_classes import BuildConfigYAML, Environment, InitConfigYAML, SystemYAML
 
 
 class ModulesAPI:
